Minor code refactors in `ui-game` module
