+++ minesweeper-ui/src/main/java/com/jayasuryat/minesweeperui/component/ZoomableContent.kt
@@ -23,7 +23,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.pointer.pointerInput
-import com.jayasuryat.minesweeperui.grid.ZoomPanState
 import kotlin.math.absoluteValue
 
 @Composable
