+++ minesweeper-ui/src/main/java/com/jayasuryat/minesweeperui/grid/Minefield.kt
@@ -15,6 +15,7 @@
  */
 package com.jayasuryat.minesweeperui.grid
 
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -22,6 +23,7 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.graphicsLayer
 import com.jayasuryat.minesweeperui.action.CellInteractionListener
+import com.jayasuryat.minesweeperui.component.ZoomPanState
 import com.jayasuryat.minesweeperui.component.ZoomableContent
 import com.jayasuryat.util.LogCompositions
 
@@ -46,20 +48,23 @@ public fun Minefield(
     val zoomPanState = remember { savableState.value }
 
     ZoomableContent(
+        modifier = modifier,
         defaultState = zoomPanState,
     ) { zoomState: ZoomPanState ->
 
         savableState.value = zoomState
 
         MineGrid(
-            modifier = modifier.then(
-                Modifier.graphicsLayer {
-                    scaleX = zoomState.scale
-                    scaleY = zoomState.scale
-                    translationX = zoomState.translationX
-                    translationY = zoomState.translationY
-                }
-            ),
+            modifier = Modifier
+                .fillMaxSize()
+                .then(
+                    Modifier.graphicsLayer {
+                        scaleX = zoomState.scale
+                        scaleY = zoomState.scale
+                        translationX = zoomState.translationX
+                        translationY = zoomState.translationY
+                    }
+                ),
             gridInfo = gridInfo,
             actionListener = actionListener,
         )
