+++ minesweeper-ui/src/main/java/com/jayasuryat/minesweeperui/theme/MinesweeperTheme.kt
@@ -23,7 +23,7 @@ import androidx.compose.runtime.staticCompositionLocalOf
 
 @Composable
 public fun MinesweeperTheme(
-    colors: MinesweeperColors = DefaultColors,
+    colors: MinesweeperColors,
     content: @Composable () -> Unit,
 ) {
